Refactor : Implementing cli with cobra library
