+++ .gitignore
@@ -1,5 +1,6 @@
 .idea
 *.iml
-cmd
+bin
 kubeslice
 kubeslice*
+.vscode
