+++ go.mod
@@ -2,4 +2,14 @@ module github.com/kubeslice/slicectl
 
 go 1.17
 
-require github.com/go-yaml/yaml v2.1.0+incompatible // indirect
+require (
+	github.com/go-yaml/yaml v2.1.0+incompatible
+	github.com/spf13/cobra v1.5.0
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/kr/pretty v0.3.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+)
